TextOFX: adding custom fonts will now override font list
See https://github.com/NatronGitHub/Natron/issues/345

Prior to this commit using custom font(s) would add them to the existing font family list, this was confusing to some users. This commit fixes this, now if you add an override ONLY the font(s) in the override are added to the font family list, everything else is removed.

NOTE! This might break older projects that uses an override and "system" fonts combined.