Text: disable fontmap usage
fontmap is not thread-safe (even though they (pango) claim it is), so disable usage.

We only use fontmap to force the usage of freetype, something that should already be default. We might need to check Windows/macOS to verify that we only use freetype. On Windows I will disable pangowin32 for Natron 2.5+